pyproject.toml
[project]
name = "docker-monitoring"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiosqlite>=0.20.0",
    "bootstrap4>=0.1.0",
    "docker>=7.1.0",
    "fastapi[standard]>=0.115.6",
    "greenlet>=3.1.1",
    "htmx>=0.0.0",
    "httpx>=0.28.1",
    "psutil>=6.1.1",
    "python-dotenv>=1.0.1",
    "python-telegram-bot>=21.10",
    "sqlalchemy>=2.0.37",
]
